cohesity-iam-scoper/requirements.txt
click>=8.0.0
rich>=13.0.0
pyyaml>=6.0
jsonschema>=4.0.0
boto3>=1.26.0
